Did some structural changes to the existing project. Made an Accordion Component.
